+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Added a debug message if no site URI could be determined from the requested URL.
 
+### Fixed
+- Fixed an edge case migration error that could occur when updating to version 3.6.6 ([#240](https://github.com/putyourlightson/craft-blitz/issues/240)).
+
 ## 3.6.6 - 2020-07-27
 ### Changed
 - Added the ability to override site URIs before they are warmed ([#230](https://github.com/putyourlightson/craft-blitz/issues/230)).